requirements.txt
google-cloud-compute
cloudshell-shell-core~=6.0
cloudshell-shell-flows~=3.0
attrs>=22,<24
typing-extensions>=4.4
